Refactor client/server scripts
Switch to `socat` for socket creation. It is installed by default on
both debian and fedora templates. It has been tested on Qubes 4.0rc2

The biggest change is the user interaction, and the security it can
provide. Each time a ssh client opens a connection to the local socket
it calls qubes.SshAgent on the target machine. This script prompts the
user to choose a SSH private key, creates a new ssh-agent and loads
just that key.

The biggest security advantages I see are:
* Ability to kill ssh-agent after 10/20/30sec timeout (Not implemented)
* Users are prompted each time a ssh client requests a key
* The client only has access to a single key file, allowing different
  keys to be used for different servers

At the moment, you only need to start the client side script:
```
user@local $ . qubes_ssh_agent
```