qubes.SshAgent: Optionally allow use of a key for a period of time
If an AppVMs key directory contains only one key file or directory,
it is now possible to specify a cache timeout in seconds in a file
called .cache next to the already existing .default file. The .cache
file must contain the number of seconds after which the user needs
to reconfirm use of a key. If the file contains -1, the user will
never get asked.